revert_jira_status: don't revert if LP hasn't actually reverted
There are some cases where the actual work will be done but the LP
status of the bug will not move, as the LP status has specific semantics
tied to a published package in the archive, and in some cases the work
will only be done in a Git repo but will wait for a later bulk upload to
be reflected in any published package.

For example, this frequently happens for systemd and glibc.

In addition, when working on an SRU, the uploaded package can be stuck
for a while in the SRU queue before publication, and can also sometimes
take a while before released into the release pocket.

Allowing a Jira issue to move to Done and stay there while the LP task
is still in one of those states makes working on those tasks much
easier, without having to remember to remove the relevant tag from the
issue.

V2:
* Fix to take a task list instead of a task
* Simplify the mocking setup code